Moved utilities into lib/ folder
