+++ lib/graham_scan.js
